댓글구현완료

+++ src/main/java/com/example/cupick/repository/BoardRepository.java
@@ -3,5 +3,5 @@ package com.example.cupick.repository;
 import com.example.cupick.model.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Long> {
+public interface BoardRepository extends JpaRepository<Board, Integer> {
 }
