+++ src/main/java/com/example/cupick/repository/ReplyRepository.java
@@ -3,5 +3,5 @@ package com.example.cupick.repository;
 import com.example.cupick.model.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyRepository extends JpaRepository<Reply, Long> {
+public interface ReplyRepository extends JpaRepository<Reply, Integer> {
 }
